renamed employee file todetailsEditContainer
